Dag1

+++ Laboration2/Entities/Borrower.cs
@@ -5,6 +5,6 @@ public class Borrower
     public int Id { get; set; }
     public string FirstName { get; set; } = string.Empty;
     public string LastName { get; set; } = string.Empty;
-    public int Ssn { get; set; }
+    public long Ssn { get; set; }
 
 }
