+++ Laboration2/Models/BookRequest.cs
@@ -3,6 +3,6 @@
 public class BookRequest
 {
     public string Title { get; set; } = string.Empty;
-    public int Isbn { get; set; }
+    public long Isbn { get; set; }
     public int ReleaseYear { get; set; }
 }
